pyproject.toml
[tool.poetry]
name = "desafio"
version = "0.1.0"
description = "Desafio do bootcamp DIO: Python com fastapi"
authors = ["Thiago Moraes <thiagovmoraess@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.135.3"
uvicorn = {extras = ["standard"], version = "^0.43.0"}
sqlalchemy = "^2.0.49"
aiosqlite = "^0.22.1"
alembic = "^1.18.4"
pydantic-settings = "^2.13.1"
pyjwt = "^2.12.1"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
email-validator = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
ruff = "^0.15.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
